feat: move log files to distinct folder
