+++ src/logger/writeToFile.ts
@@ -1,6 +1,9 @@
 const fs = require('fs');
 
-const writeToFile = (text: string, file='./logs/errors.log') => {
+const writeToFile = (text: string, file='./logs/errors.log', folder='./logs') => {
+  if (!fs.existsSync(folder)) {
+    fs.mkdirSync(folder);
+  }
   fs.appendFileSync(
     file,
     `${text} \r\n`,
